+++ src/main/java/org/opentripplanner/graph_builder/linking/StreetSplitter.java
@@ -1,6 +1,9 @@
 package org.opentripplanner.graph_builder.linking;
 
 import com.google.common.collect.Iterables;
+import gnu.trove.map.TIntDoubleMap;
+import gnu.trove.map.hash.TIntDoubleHashMap;
+import jersey.repackaged.com.google.common.collect.Lists;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.GeometryFactory;
@@ -8,9 +11,6 @@ import org.locationtech.jts.geom.LineString;
 import org.locationtech.jts.index.SpatialIndex;
 import org.locationtech.jts.linearref.LinearLocation;
 import org.locationtech.jts.linearref.LocationIndexedLine;
-import gnu.trove.map.TIntDoubleMap;
-import gnu.trove.map.hash.TIntDoubleHashMap;
-import jersey.repackaged.com.google.common.collect.Lists;
 import org.opentripplanner.common.geometry.GeometryUtils;
 import org.opentripplanner.common.geometry.HashGridSpatialIndex;
 import org.opentripplanner.common.geometry.SphericalDistanceLibrary;
@@ -18,34 +18,36 @@ import org.opentripplanner.common.model.GenericLocation;
 import org.opentripplanner.common.model.P2;
 import org.opentripplanner.graph_builder.annotation.BikeParkUnlinked;
 import org.opentripplanner.graph_builder.annotation.BikeRentalStationUnlinked;
+import org.opentripplanner.graph_builder.annotation.StopLinkedTooFar;
 import org.opentripplanner.graph_builder.annotation.StopUnlinked;
 import org.opentripplanner.graph_builder.services.DefaultStreetEdgeFactory;
 import org.opentripplanner.graph_builder.services.StreetEdgeFactory;
 import org.opentripplanner.openstreetmap.model.OSMWithTags;
-import org.opentripplanner.graph_builder.annotation.StopLinkedTooFar;
 import org.opentripplanner.routing.core.RoutingRequest;
 import org.opentripplanner.routing.core.TraverseMode;
 import org.opentripplanner.routing.core.TraverseModeSet;
+import org.opentripplanner.routing.edgetype.AreaEdge;
+import org.opentripplanner.routing.edgetype.AreaEdgeList;
+import org.opentripplanner.routing.edgetype.SemiPermanentPartialStreetEdge;
 import org.opentripplanner.routing.edgetype.StreetBikeParkLink;
 import org.opentripplanner.routing.edgetype.StreetBikeRentalLink;
 import org.opentripplanner.routing.edgetype.StreetEdge;
 import org.opentripplanner.routing.edgetype.StreetTransitLink;
-import org.opentripplanner.routing.edgetype.TemporaryFreeEdge;
-import org.opentripplanner.routing.edgetype.AreaEdgeList;
-import org.opentripplanner.routing.edgetype.AreaEdge;
 import org.opentripplanner.routing.edgetype.StreetTraversalPermission;
+import org.opentripplanner.routing.edgetype.TemporaryFreeEdge;
 import org.opentripplanner.routing.graph.Edge;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.Vertex;
 import org.opentripplanner.routing.location.TemporaryStreetLocation;
 import org.opentripplanner.routing.vertextype.BikeParkVertex;
 import org.opentripplanner.routing.vertextype.BikeRentalStationVertex;
+import org.opentripplanner.routing.vertextype.IntersectionVertex;
+import org.opentripplanner.routing.vertextype.SemiPermanentSplitterVertex;
 import org.opentripplanner.routing.vertextype.SplitterVertex;
 import org.opentripplanner.routing.vertextype.StreetVertex;
 import org.opentripplanner.routing.vertextype.TemporarySplitterVertex;
 import org.opentripplanner.routing.vertextype.TemporaryVertex;
 import org.opentripplanner.routing.vertextype.TransitStop;
-import org.opentripplanner.routing.vertextype.IntersectionVertex;
 import org.opentripplanner.util.I18NString;
 import org.opentripplanner.util.LocalizedString;
 import org.opentripplanner.util.NonLocalizedString;
@@ -54,7 +56,9 @@ import org.slf4j.LoggerFactory;
 
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -68,22 +72,35 @@ import java.util.stream.Collectors;
  * iterate over Maps and even in the presence of points that are exactly halfway between multiple candidate linking
  * points.
  *
- * It would be wise to keep this new incarnation of the linking code relatively simple, considering what happened before.
- *
  * See discussion in pull request #1922, follow up issue #1934, and the original issue calling for replacement of
  * the stop linker, #1305.
  */
-public class SimpleStreetSplitter {
+public class StreetSplitter {
 
-    private static final Logger LOG = LoggerFactory.getLogger(SimpleStreetSplitter.class);
+    private static final Logger LOG = LoggerFactory.getLogger(StreetSplitter.class);
 
     public static final int MAX_SEARCH_RADIUS_METERS = 1000;
 
     private Boolean addExtraEdgesToAreas = false;
 
-    private StreetEdgeFactory edgeFactory;
+    private StreetEdgeFactory edgeFactory = new DefaultStreetEdgeFactory();
 
     public static final int WARNING_DISTANCE_METERS = 20;
+
+    /**
+     * DESTRUCTIVE_SPLIT means edges are split and new edges and vertices are created. This should occur when linking
+     * transit stops etc. during graph building or inserting floating bicycles in a bikeshare system in realtime into
+     * the graph. When these splits occur, a spatial index of all edges should be kept updated, otherwise there will
+     * appear to be snapping issues as observed in https://github.com/opentripplanner/OpenTripPlanner/issues/2758.
+     */
+    public static final boolean DESTRUCTIVE_SPLIT = true;
+
+    /**
+     * NON_DESTRUCTIVE_SPLIT means new temporary edges and vertices are created and no edges are deleted. This is used
+     * in individual requests when temporarily inserting the origin/destination into the graph. The spatial index of
+     * street edges should not be updated to include these temporary edges.
+     */
+    public static final boolean NON_DESTRUCTIVE_SPLIT = false;
 
     /** if there are two ways and the distances to them differ by less than this value, we link to both of them */
     public static final double DUPLICATE_WAY_EPSILON_METERS = 0.001;
@@ -96,31 +113,41 @@ public class SimpleStreetSplitter {
 
     private static GeometryFactory geometryFactory = GeometryUtils.getGeometryFactory();
 
-    //If true edges are split and new edges are created (used when linking transit stops etc. during graph building)
-    //If false new temporary edges are created and no edges are deleted (Used when searching for origin/destination)
-    private final boolean destructiveSplitting;
+    // used to keep track of splitters associated with graphs to make sure that
+    public static final HashSet<Graph> graphsWithSplitters = new HashSet<>();
 
     /**
-     * Construct a new SimpleStreetSplitter.
-     * NOTE: Only one SimpleStreetSplitter should be active on a graph at any given time.
+     * Construct a new StreetSplitter.
+     *
+     * IMPORTANT NOTE: Only one StreetSplitter should be active on a graph at any given time. This is because any
+     * newly-created split edges in this street splitter won't show up in the index of the other street splitters
+     * (and potentially StreetVertexIndexServices).
      *
      * @param hashGridSpatialIndex If not null this index is used instead of creating new one
      * @param transitStopIndex Index of all transitStops which is generated in {@link org.opentripplanner.routing.impl.StreetVertexIndexServiceImpl}
-     * @param destructiveSplitting If true splitting is permanent (Used when linking transit stops etc.) when false Splitting is only for duration of a request. Since they are made from temporary vertices and edges.
      */
-    public SimpleStreetSplitter(Graph graph, HashGridSpatialIndex<Edge> hashGridSpatialIndex,
-        SpatialIndex transitStopIndex, boolean destructiveSplitting) {
+    public StreetSplitter(
+        Graph graph,
+        HashGridSpatialIndex<Edge> hashGridSpatialIndex,
+        SpatialIndex transitStopIndex
+    ) {
         this.graph = graph;
+        if (graphsWithSplitters.contains(graph)) {
+            LOG.warn("Multiple StreetSplitters detected on the same graph! Make sure only the most recently created one is used going forward!");
+        }
+        LOG.info("New StreetSpiltter created successfully!");
+        graphsWithSplitters.add(graph);
         this.transitStopIndex = transitStopIndex;
-        this.destructiveSplitting = destructiveSplitting;
-        this.edgeFactory = new DefaultStreetEdgeFactory();
 
         //We build a spatial index if it isn't provided
         if (hashGridSpatialIndex == null) {
             // build a nice private spatial index, since we're adding and removing edges
-            idx = new HashGridSpatialIndex<>();
-            for (StreetEdge se : Iterables.filter(graph.getEdges(), StreetEdge.class)) {
-                idx.insert(se.getGeometry(), se);
+            // place a lock while doing this to ensure thread-safe writes to the index
+            synchronized (this) {
+                idx = new HashGridSpatialIndex<>();
+                for (StreetEdge se : Iterables.filter(graph.getEdges(), StreetEdge.class)) {
+                    idx.insert(se.getGeometry(), se);
+                }
             }
         } else {
             idx = hashGridSpatialIndex;
@@ -129,21 +156,25 @@ public class SimpleStreetSplitter {
     }
 
     /**
-     * Construct a new SimpleStreetSplitter. Be aware that only one SimpleStreetSplitter should be
+     * Construct a new StreetSplitter. Be aware that only one StreetSplitter should be
      * active on a graph at any given time.
      *
-     * SimpleStreetSplitter generates index on graph and splits destructively (used in transit splitter)
+     * StreetSplitter generates index on graph and splits destructively (used in transit splitter)
      * @param graph
      */
-    public SimpleStreetSplitter(Graph graph) {
-        this(graph, null, null, true);
+    public StreetSplitter(Graph graph) {
+        this(graph, null, null);
     }
 
-    /** Link all relevant vertices to the street network */
-    public void link () {	
+    /**
+     * Link all relevant vertices to the street network.
+     * This is to be called only during graph building in order to destructively split edges in order to create
+     * permanent edges.
+     */
+    public void linkAllStationsToGraph() {
         for (Vertex v : graph.getVertices()) {
             if (v instanceof TransitStop || v instanceof BikeRentalStationVertex || v instanceof BikeParkVertex)
-                if (!link(v)) {
+                if (!linkToClosestWalkableEdge(v, DESTRUCTIVE_SPLIT, false)) {
                     if (v instanceof TransitStop)
                         LOG.warn(graph.addBuilderAnnotation(new StopUnlinked((TransitStop) v)));
                     else if (v instanceof BikeRentalStationVertex)
@@ -154,13 +185,69 @@ public class SimpleStreetSplitter {
         }
     }
 
-    /** Link this vertex into the graph to the closest walkable edge */
-    public boolean link (Vertex vertex) {
-        return link(vertex, TraverseMode.WALK, null);
+    /**
+     * Link this vertex into the graph to the closest walkable edge.
+     *
+     * @param vertex The vertex to be linked.
+     * @param destructiveSplitting If true splitting is permanent (Used when linking transit stops etc.) when
+     *                             false the original edge is kept in the graph after creating split edges.
+     * @param createSemiPermanentEdges If true, semi-permanent edges and a semi-permanent splitter vertex is created.
+     *                                 These vertices and edges are intended to be kept in the graph and linked to
+     *                                 during subsequent requests, but might be removed at a later time. For example a
+     *                                 floating bike rental may need to be removed from the graph once the data is
+     *                                 updated to show that the floating bike rental no longer exists or if its
+     *                                 location has significantly changed.
+     */
+    public boolean linkToClosestWalkableEdge(
+        Vertex vertex,
+        final boolean destructiveSplitting,
+        boolean createSemiPermanentEdges
+    ) {
+        return linkToGraph(vertex, TraverseMode.WALK, null, destructiveSplitting, createSemiPermanentEdges);
     }
 
-    /** Link this vertex into the graph */
-    public boolean link(Vertex vertex, TraverseMode traverseMode, RoutingRequest options) {
+    /**
+     * Link a vertex into the graph to the closest edge that allows the traversal of the specified mode.
+     */
+    public boolean linkToGraph(
+        Vertex vertex,
+        TraverseMode traverseMode,
+        RoutingRequest options,
+        final boolean destructiveSplitting,
+        boolean createSemiPermanentEdges
+    ) {
+        final TraverseModeSet traverseModeSet = new TraverseModeSet(traverseMode);
+        // also set the walk mode to true for bicycling. This is done because it is assumed that the bike is able to be
+        // walked along certain edges that explicitly forbid bicycle travel.
+        if (traverseMode == TraverseMode.BICYCLE) {
+            traverseModeSet.setWalk(true);
+        }
+        return linkToGraph(vertex, traverseModeSet, options, destructiveSplitting, createSemiPermanentEdges);
+    }
+
+    /**
+     *  Link this vertex into the graph (ie, connect the vertex to other vertices in the graph by creating one or more
+     *  new edges/vertexes which are then connected to the other vertices/edges of the graph.
+     *
+     * @param vertex The vertex to be linked.
+     * @param traverseModeSet The traverse modes.
+     * @param options The routing options.
+     * @param destructiveSplitting If true, splitting is permanent (Used when linking transit stops etc). When
+     *                             false the original edge is kept in the graph after creating split edges.
+     * @param createSemiPermanentEdges If true, semi-permanent edges and a semi-permanent splitter vertex is created.
+     *                                 These vertices and edges are intended to be kept in the graph and linked to
+     *                                 during subsequent requests, but might be removed at a later time. For example a
+     *                                 floating bike rental may need to be removed from the graph once the data is
+     *                                 updated to show that the floating bike rental no longer exists or if its
+     *                                 location has significantly changed.
+     */
+    public boolean linkToGraph(
+        Vertex vertex,
+        TraverseModeSet traverseModeSet,
+        RoutingRequest options,
+        final boolean destructiveSplitting,
+        boolean createSemiPermanentEdges
+    ) {
         // find nearby street edges
         // TODO: we used to use an expanding-envelope search, which is more efficient in
         // dense areas. but first let's see how inefficient this is. I suspect it's not too
@@ -177,18 +264,17 @@ public class SimpleStreetSplitter {
 
         final double DUPLICATE_WAY_EPSILON_DEGREES = SphericalDistanceLibrary.metersToDegrees(DUPLICATE_WAY_EPSILON_METERS);
 
-        final TraverseModeSet traverseModeSet;
-        if (traverseMode == TraverseMode.BICYCLE) {
-            traverseModeSet = new TraverseModeSet(traverseMode, TraverseMode.WALK);
-        } else {
-            traverseModeSet = new TraverseModeSet(traverseMode);
-        }
         // We sort the list of candidate edges by distance to the stop
         // This should remove any issues with things coming out of the spatial index in different orders
-        // Then we link to everything that is within DUPLICATE_WAY_EPSILON_METERS of of the best distance
+        // Then we link to everything that is within DUPLICATE_WAY_EPSILON_METERS of the best distance
         // so that we capture back edges and duplicate ways.
         List<StreetEdge> candidateEdges = idx.query(env).stream()
-            .filter(streetEdge -> streetEdge instanceof  StreetEdge)
+            .filter(
+                streetEdge -> streetEdge instanceof StreetEdge &&
+                    // Do not find SemiPermanentPartialStreetEdges if creating new SemiPermanentPartialStreetEdges.
+                    // Use only permanent StreetEdges instead
+                    (!createSemiPermanentEdges || !(streetEdge instanceof SemiPermanentPartialStreetEdge))
+            )
             .map(edge -> (StreetEdge) edge)
             // note: not filtering by radius here as distance calculation is expensive
             // we do that below.
@@ -207,10 +293,8 @@ public class SimpleStreetSplitter {
         Collections.sort(candidateEdges, (o1, o2) -> {
             double diff = distances.get(o1.getId()) - distances.get(o2.getId());
             // A Comparator must return an integer but our distances are doubles.
-            if (diff < 0)
-                return -1;
-            if (diff > 0)
-                return 1;
+            if (diff < 0) return -1;
+            if (diff > 0) return 1;
             return 0;
         });
 
@@ -236,12 +320,8 @@ public class SimpleStreetSplitter {
 
             Collections.sort(candidateStops, (o1, o2) -> {
                     double diff = stopDistances.get(o1.getIndex()) - stopDistances.get(o2.getIndex());
-                    if (diff < 0) {
-                        return -1;
-                    }
-                    if (diff > 0) {
-                        return 1;
-                    }
+                    if (diff < 0) return -1;
+                    if (diff > 0) return 1;
                     return 0;
             });
             if (candidateStops.isEmpty() || stopDistances.get(candidateStops.get(0).getIndex()) > radiusDeg) {
@@ -262,7 +342,7 @@ public class SimpleStreetSplitter {
 
                 for (TransitStop stop: bestStops) {
                     LOG.debug("Linking vertex to stop: {}", stop.getName());
-                    makeTemporaryEdges((TemporaryStreetLocation)vertex, stop);
+                    makeTemporaryEdges((TemporaryStreetLocation)vertex, stop, destructiveSplitting);
                 }
                 return true;
             }
@@ -283,7 +363,7 @@ public class SimpleStreetSplitter {
                     .get(candidateEdges.get(i - 1).getId()) < DUPLICATE_WAY_EPSILON_DEGREES);
 
             for (StreetEdge edge : bestEdges) {
-                link(vertex, edge, xscale, options);
+                linkToEdge(vertex, edge, xscale, options, destructiveSplitting, createSemiPermanentEdges);
             }
 
             // Warn if a linkage was made, but the linkage was suspiciously long.
@@ -300,7 +380,9 @@ public class SimpleStreetSplitter {
         }
     }
 
-    // Link to all vertices in area/platform
+    /**
+     * Link to all vertices in area/platform
+     */
     private void linkTransitToAreaVertices(Vertex splitterVertex, AreaEdgeList area) {
         List<Vertex> vertices = new ArrayList<>();
 
@@ -310,7 +392,7 @@ public class SimpleStreetSplitter {
         }
 
         for (Vertex vertex : vertices) {
-            if (vertex instanceof  StreetVertex && !vertex.equals(splitterVertex)) {
+            if (vertex instanceof StreetVertex && !vertex.equals(splitterVertex)) {
                 LineString line = geometryFactory.createLineString(new Coordinate[] { splitterVertex.getCoordinate(), vertex.getCoordinate()});
                 double length = SphericalDistanceLibrary.distance(splitterVertex.getCoordinate(),
                         vertex.getCoordinate());
@@ -322,38 +404,59 @@ public class SimpleStreetSplitter {
         }
     }
 
-    /** split the edge and link in the transit stop */
-    private void link(Vertex tstop, StreetEdge edge, double xscale, RoutingRequest options) {
+    /**
+     * Split the edge and link in the transit stop.
+     * @param vertex An object of Vertex to be linked to an edge.
+     * @param edge An object of StreetEdge to be linked to.
+     * @param xscale The longitude scale factor in Equirectangular projection.
+     * @param options An object of RoutingRequest
+     * @param destructiveSplitting If true splitting is permanent (Used when linking transit stops etc.) when
+     *                             false the original edge is kept in the graph after creating split edges.
+     * @param createSemiPermanentEdges If true, semi-permanent edges and a semi-permanent splitter vertex is created.
+     *                                 These vertices and edges are intended to be kept in the graph and linked to
+     *                                 during subsequent requests, but might be removed at a later time. For example a
+     *                                 floating bike rental may need to be removed from the graph once the data is
+     *                                 updated to show that the floating bike rental no longer exists or if its
+     *                                 location has significantly changed.
+     */
+    private void linkToEdge(
+        Vertex vertex,
+        StreetEdge edge,
+        double xscale,
+        RoutingRequest options,
+        final boolean destructiveSplitting,
+        boolean createSemiPermanentEdges
+    ) {
         // TODO: we've already built this line string, we should save it
         LineString orig = edge.getGeometry();
         LineString transformed = equirectangularProject(orig, xscale);
         LocationIndexedLine il = new LocationIndexedLine(transformed);
-        LinearLocation ll = il.project(new Coordinate(tstop.getLon() * xscale, tstop.getLat()));
+        LinearLocation ll = il.project(new Coordinate(vertex.getLon() * xscale, vertex.getLat()));
 
         // if we're very close to one end of the line or the other, or endwise, don't bother to split,
         // cut to the chase and link directly
         // We use a really tiny epsilon here because we only want points that actually snap to exactly the same location on the
         // street to use the same vertices. Otherwise the order the stops are loaded in will affect where they are snapped.
         if (ll.getSegmentIndex() == 0 && ll.getSegmentFraction() < 1e-8) {
-            makeLinkEdges(tstop, (StreetVertex) edge.getFromVertex());
+            makeLinkEdges(vertex, (StreetVertex) edge.getFromVertex(), destructiveSplitting);
         }
         // -1 converts from count to index. Because of the fencepost problem, npoints - 1 is the "segment"
         // past the last point
         else if (ll.getSegmentIndex() == orig.getNumPoints() - 1) {
-            makeLinkEdges(tstop, (StreetVertex) edge.getToVertex());
+            makeLinkEdges(vertex, (StreetVertex) edge.getToVertex(), destructiveSplitting);
         }
 
         // nPoints - 2: -1 to correct for index vs count, -1 to account for fencepost problem
         else if (ll.getSegmentIndex() == orig.getNumPoints() - 2 && ll.getSegmentFraction() > 1 - 1e-8) {
-            makeLinkEdges(tstop, (StreetVertex) edge.getToVertex());
+            makeLinkEdges(vertex, (StreetVertex) edge.getToVertex(), destructiveSplitting);
         }
 
         else {
 
             TemporaryVertex temporaryVertex = null;
             boolean endVertex = false;
-            if (tstop instanceof TemporaryVertex) {
-                temporaryVertex = (TemporaryVertex) tstop;
+            if (vertex instanceof TemporaryVertex) {
+                temporaryVertex = (TemporaryVertex) vertex;
                 endVertex = temporaryVertex.isEndVertex();
 
             }
@@ -363,12 +466,34 @@ public class SimpleStreetSplitter {
                 options.canSplitEdge(edge);
             }
             // split the edge, get the split vertex
-            SplitterVertex v0 = split(edge, ll, temporaryVertex != null, endVertex);
-            makeLinkEdges(tstop, v0);
+            SplitterVertex v0 = split(
+                edge,
+                ll,
+                temporaryVertex != null,
+                endVertex,
+                destructiveSplitting,
+                createSemiPermanentEdges
+            );
+
+            // if SemiPermanentEdges were created, update some of their characteristics to ensure that they are not
+            // traversed if there is no reason to do so
+            if (createSemiPermanentEdges) {
+                Set<Edge> newlySplitEdges = new HashSet<>();
+                newlySplitEdges.addAll(v0.getOutgoing());
+                newlySplitEdges.addAll(v0.getIncoming());
+
+                if (vertex instanceof BikeRentalStationVertex) {
+                    for (Edge splitEdge : newlySplitEdges) {
+                        ((SemiPermanentPartialStreetEdge) splitEdge).setBikeRentalOptionRequired();
+                    }
+                }
+            }
+
+            makeLinkEdges(vertex, v0, destructiveSplitting);
 
             // If splitter vertex is part of area; link splittervertex to all other vertexes in area, this creates
             // edges that were missed by WalkableAreaBuilder
-            if (edge instanceof AreaEdge && tstop instanceof TransitStop && this.addExtraEdgesToAreas) {
+            if (edge instanceof AreaEdge && vertex instanceof TransitStop && this.addExtraEdgesToAreas) {
                 linkTransitToAreaVertices(v0, ((AreaEdge) edge).getArea());
             }
         }
@@ -380,11 +505,28 @@ public class SimpleStreetSplitter {
      *
      * @param edge to be split
      * @param ll fraction at which to split the edge
-     * @param temporarySplit if true this is temporary split at origin/destinations search and only temporary edges vertices are created
+     * @param temporarySplit if true this is temporary split at origin/destinations search and only temporary edges
+     *                       vertices are created
      * @param endVertex if this is temporary edge this is true if this is end vertex otherwise it doesn't matter
+     * @param destructiveSplitting If true splitting is permanent (Used when linking transit stops etc.) when
+     *                             false the original edge is kept in the graph after creating split edges.
+     * @param createSemiPermanentEdges If true, semi-permanent edges and a semi-permanent splitter vertex is created.
+     *                                 These vertices and edges are intended to be kept in the graph and linked to
+     *                                 during subsequent requests, but might be removed at a later time. For example a
+     *                                 floating bike rental may need to be removed from the graph once the data is
+     *                                 updated to show that the floating bike rental no longer exists or if its
+     *                                 location has significantly changed.
      * @return Splitter vertex with added new edges
      */
-    private SplitterVertex split (StreetEdge edge, LinearLocation ll, boolean temporarySplit, boolean endVertex) {
+    private SplitterVertex split (
+        StreetEdge edge,
+        LinearLocation ll,
+        boolean temporarySplit,
+        boolean endVertex,
+        final boolean destructiveSplitting,
+        boolean createSemiPermanentEdges
+    ) {
+
         LineString geometry = edge.getGeometry();
 
         // create the geometries
@@ -392,24 +534,45 @@ public class SimpleStreetSplitter {
 
         // every edge can be split exactly once, so this is a valid label
         SplitterVertex v;
-        if (temporarySplit) {
-            v = new TemporarySplitterVertex("split from " + edge.getId(), splitPoint.x, splitPoint.y, edge, endVertex);
+        StringBuilder splitLabel = new StringBuilder();
+        splitLabel.append("split from ");
+        splitLabel.append(edge.getId());
+        splitLabel.append(" #");
+        int splintNum = 1;
+        splitLabel.append(splintNum);
+        while (graph.containsVertexLabel(splitLabel.toString())) {
+            splintNum++;
+            splitLabel.replace(splitLabel.lastIndexOf("#") + 1, Integer.MAX_VALUE, String.valueOf(splintNum));
+        }
+        String vertexLabel = splitLabel.toString();
+        if (createSemiPermanentEdges) {
+            v = new SemiPermanentSplitterVertex(graph, vertexLabel, splitPoint.x, splitPoint.y, edge);
+        } else if (temporarySplit) {
+            v = new TemporarySplitterVertex(vertexLabel, splitPoint.x, splitPoint.y, edge, endVertex);
         } else {
-            v = new SplitterVertex(graph, "split from " + edge.getId(), splitPoint.x, splitPoint.y, edge);
+            v = new SplitterVertex(graph, vertexLabel, splitPoint.x, splitPoint.y, edge);
         }
 
         // Split the 'edge' at 'v' in 2 new edges and connect these 2 edges to the
         // existing vertices
-        P2<StreetEdge> edges = edge.split(v, !temporarySplit);
+        P2<StreetEdge> edges = edge.split(v, destructiveSplitting, createSemiPermanentEdges);
+
+        if (destructiveSplitting || createSemiPermanentEdges) {
+            // update indices of new edges
+            synchronized (this) {
+                // Note: Write operations are not synchronized in HashGridSpatialIndex, hence the lock.
+                idx.insert(edges.first.getGeometry(), edges.first);
+                idx.insert(edges.second.getGeometry(), edges.second);
+
+                // if destructively splitting, remove the edge from the index
+                if (destructiveSplitting) {
+                    idx.remove(edge.getGeometry().getEnvelopeInternal(), edge);
+                }
+            }
+        }
 
         if (destructiveSplitting) {
-            // update indices of new edges
-            idx.insert(edges.first.getGeometry(), edges.first);
-            idx.insert(edges.second.getGeometry(), edges.second);
-
-            // (no need to remove original edge, we filter it when it comes out of the index)
-
-            // remove original edge from the graph
+            // remove original edge from the graph if split was destructive
             edge.getToVertex().removeIncoming(edge);
             edge.getFromVertex().removeOutgoing(edge);
         }
@@ -418,37 +581,48 @@ public class SimpleStreetSplitter {
     }
 
     /** Make the appropriate type of link edges from a vertex */
-    private void makeLinkEdges(Vertex from, StreetVertex to) {
+    private void makeLinkEdges(Vertex from, StreetVertex to, final boolean destructiveSplitting) {
         if (from instanceof TemporaryStreetLocation) {
-            makeTemporaryEdges((TemporaryStreetLocation) from, to);
+            makeTemporaryEdges((TemporaryStreetLocation) from, to, destructiveSplitting);
         } else if (from instanceof TransitStop) {
-            makeTransitLinkEdges((TransitStop) from, to);
+            makeTransitLinkEdges((TransitStop) from, to, destructiveSplitting);
         } else if (from instanceof BikeRentalStationVertex) {
             makeBikeRentalLinkEdges((BikeRentalStationVertex) from, to);
         } else if (from instanceof BikeParkVertex) {
-            makeBikeParkEdges((BikeParkVertex) from, to);
+            makeBikeParkEdges((BikeParkVertex) from, to, destructiveSplitting);
         }
     }
 
-    /** Make temporary edges to origin/destination vertex in origin/destination search **/
-    private void makeTemporaryEdges(TemporaryStreetLocation from, Vertex to) {
+    /**
+     * Make temporary edges to origin/destination vertex in origin/destination search
+     */
+    private void makeTemporaryEdges(TemporaryStreetLocation from, Vertex to, final boolean destructiveSplitting) {
         if (destructiveSplitting) {
             throw new RuntimeException("Destructive splitting is used on temporary edges. Something is wrong!");
         }
         if (to instanceof TemporarySplitterVertex) {
             from.setWheelchairAccessible(((TemporarySplitterVertex) to).isWheelchairAccessible());
         }
+        // there is no need to create a TemporaryFreeEdge when the to vertex
         if (from.isEndVertex()) {
+            // Don't create a TemporaryFreeEdge if the to vertex doesn't have any incoming edges, it's a dead end that
+            // was probably caused by avoiding the creation of identical TemporaryPartialStreetEdges in cases where a
+            // certain SemiPermanentPartialStreetEdge was split.
+            if (to.getDegreeIn() == 0) return;
             LOG.debug("Linking end vertex to {} -> {}", to, from);
             new TemporaryFreeEdge(to, from);
         } else {
+            // Don't create a TemporaryFreeEdge if the to vertex doesn't have any outgoing edges, it's a dead end that
+            // was probably caused by avoiding the creation of identical TemporaryPartialStreetEdges in cases where a
+            // certain SemiPermanentPartialStreetEdge was split.
+            if (to.getDegreeOut() == 0) return;
             LOG.debug("Linking start vertex to {} -> {}", from, to);
             new TemporaryFreeEdge(from, to);
         }
     }
 
     /** Make bike park edges */
-    private void makeBikeParkEdges(BikeParkVertex from, StreetVertex to) {
+    private void makeBikeParkEdges(BikeParkVertex from, StreetVertex to, final boolean destructiveSplitting) {
         if (!destructiveSplitting) {
             throw new RuntimeException("Bike park edges are created with non destructive splitting!");
         }
@@ -461,15 +635,16 @@ public class SimpleStreetSplitter {
         new StreetBikeParkLink(to, from);
     }
 
-    /** 
+    /**
      * Make street transit link edges, unless they already exist.
      */
-    private void makeTransitLinkEdges (TransitStop tstop, StreetVertex v) {
+    private void makeTransitLinkEdges (TransitStop tstop, StreetVertex v, final boolean destructiveSplitting) {
         if (!destructiveSplitting) {
             throw new RuntimeException("Transitedges are created with non destructive splitting!");
         }
         // ensure that the requisite edges do not already exist
-        // this can happen if we link to duplicate ways that have the same start/end vertices.
+        // this can happen if we link to duplicate ways that have the same start/end vertices or if
+        // certain transit stops were already linked to the network in the TransitToTaggedStopsModule.
         for (StreetTransitLink e : Iterables.filter(tstop.getOutgoing(), StreetTransitLink.class)) {
             if (e.getToVertex() == v)
                 return;
@@ -481,9 +656,6 @@ public class SimpleStreetSplitter {
 
     /** Make link edges for bike rental */
     private void makeBikeRentalLinkEdges (BikeRentalStationVertex from, StreetVertex to) {
-        if (!destructiveSplitting) {
-            throw new RuntimeException("Bike rental edges are created with non destructive splitting!");
-        }
         for (StreetBikeRentalLink sbrl : Iterables.filter(from.getOutgoing(), StreetBikeRentalLink.class)) {
             if (sbrl.getToVertex() == to)
                 return;
@@ -526,18 +698,15 @@ public class SimpleStreetSplitter {
      *
      * Split edges don't replace existing ones and only temporary edges and vertices are created.
      *
-     * Will throw ThrivialPathException if origin and destination Location are on the same edge
+     * Will throw TrivialPathException if origin and destination Location are on the same edge
      *
      * @param location
      * @param options
      * @param endVertex true if this is destination vertex
      * @return
      */
-    public Vertex getClosestVertex(GenericLocation location, RoutingRequest options,
-        boolean endVertex) {
-        if (destructiveSplitting) {
-            throw new RuntimeException("Origin and destination search is used with destructive splitting. Something is wrong!");
-        }
+    public Vertex linkOriginDestination(GenericLocation location, RoutingRequest options, boolean endVertex) {
+
         if (endVertex) {
             LOG.debug("Finding end vertex for {}", location);
         } else {
@@ -576,7 +745,7 @@ public class SimpleStreetSplitter {
                 nonTransitMode = TraverseMode.BICYCLE;
         }
 
-        if(!link(closest, nonTransitMode, options)) {
+        if(!linkToGraph(closest, nonTransitMode, options, NON_DESTRUCTIVE_SPLIT, false)) {
             LOG.warn("Couldn't link {}", location);
         }
         return closest;
@@ -589,5 +758,29 @@ public class SimpleStreetSplitter {
 
     public void setAddExtraEdgesToAreas(Boolean addExtraEdgesToAreas) {
         this.addExtraEdgesToAreas = addExtraEdgesToAreas;
+    }
+
+    public void removeSemiPermanentVerticesAndEdges (SemiPermanentSplitterVertex vertex) {
+        // remove all associated SemiPermanentPartialStreetEdges from the splitter index
+        for (Edge edge : vertex.getOutgoing()) {
+            if (edge instanceof SemiPermanentPartialStreetEdge) {
+                removeEdgeFromIdx((StreetEdge) edge);
+            }
+        }
+        for (Edge edge : vertex.getIncoming()) {
+            if (edge instanceof SemiPermanentPartialStreetEdge) {
+                removeEdgeFromIdx((StreetEdge) edge);
+            }
+        }
+        if (graph.containsVertex(vertex)) {
+            // remove the SemiPermanentSplitterVertex and associated SemiPermanentPartialStreetEdges
+            graph.removeVertexAndEdges(vertex);
+        }
+    }
+
+    public void removeEdgeFromIdx (StreetEdge edge) {
+        synchronized (this) {
+            idx.remove(edge.getGeometry().getEnvelopeInternal(), edge);
+        }
     }
 }