+++ src/main/java/org/opentripplanner/graph_builder/module/TransitToTaggedStopsModule.java
@@ -8,7 +8,7 @@ import org.opentripplanner.routing.edgetype.StreetTransitLink;
 import org.opentripplanner.routing.graph.Edge;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.Vertex;
-import org.opentripplanner.routing.impl.StreetVertexIndexServiceImpl;
+import org.opentripplanner.routing.services.StreetVertexIndexService;
 import org.opentripplanner.routing.vertextype.TransitStop;
 import org.opentripplanner.routing.vertextype.TransitStopStreetVertex;
 import org.slf4j.Logger;
@@ -36,7 +36,7 @@ public class TransitToTaggedStopsModule implements GraphBuilderModule {
 
     private static final Logger LOG = LoggerFactory.getLogger(TransitToTaggedStopsModule.class);
 
-    StreetVertexIndexServiceImpl index;
+    StreetVertexIndexService index;
     private double searchRadiusM = 250;
     private double searchRadiusLat = SphericalDistanceLibrary.metersToDegrees(searchRadiusM);
 
@@ -52,7 +52,11 @@ public class TransitToTaggedStopsModule implements GraphBuilderModule {
     public void buildGraph(Graph graph, HashMap<Class<?>, Object> extra) {
         LOG.info("Linking transit stops to tagged bus stops...");
 
-        index = new StreetVertexIndexServiceImpl(graph);
+        // Make sure the graph index has been initialized. Don't recalculate the street index because it should
+        // already have been initialized in a previous build module and should have all needed vertex data for
+        // spatially querying vertices.
+        graph.index(false);
+        index = graph.streetIndex;
 
         // iterate over a copy of vertex list because it will be modified
         ArrayList<Vertex> vertices = new ArrayList<>();