+++ src/main/java/org/opentripplanner/updater/bike_rental/BikeRentalUpdater.java
@@ -1,5 +1,26 @@
 package org.opentripplanner.updater.bike_rental;
 
+import com.fasterxml.jackson.databind.JsonNode;
+import org.opentripplanner.graph_builder.linking.StreetSplitter;
+import org.opentripplanner.routing.bike_rental.BikeRentalStation;
+import org.opentripplanner.routing.bike_rental.BikeRentalStationService;
+import org.opentripplanner.routing.edgetype.RentABikeOffEdge;
+import org.opentripplanner.routing.edgetype.RentABikeOnEdge;
+import org.opentripplanner.routing.edgetype.SemiPermanentPartialStreetEdge;
+import org.opentripplanner.routing.edgetype.StreetBikeRentalLink;
+import org.opentripplanner.routing.edgetype.StreetEdge;
+import org.opentripplanner.routing.graph.Edge;
+import org.opentripplanner.routing.graph.Graph;
+import org.opentripplanner.routing.vertextype.BikeRentalStationVertex;
+import org.opentripplanner.routing.vertextype.SemiPermanentSplitterVertex;
+import org.opentripplanner.routing.vertextype.StreetVertex;
+import org.opentripplanner.updater.GraphUpdaterManager;
+import org.opentripplanner.updater.GraphWriterRunnable;
+import org.opentripplanner.updater.JsonConfigurable;
+import org.opentripplanner.updater.PollingGraphUpdater;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -9,22 +30,9 @@ import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
 import java.util.concurrent.ExecutionException;
-import com.fasterxml.jackson.databind.JsonNode;
-import org.opentripplanner.graph_builder.linking.SimpleStreetSplitter;
-import org.opentripplanner.routing.bike_rental.BikeRentalStation;
-import org.opentripplanner.routing.bike_rental.BikeRentalStationService;
-import org.opentripplanner.routing.edgetype.RentABikeOffEdge;
-import org.opentripplanner.routing.edgetype.RentABikeOnEdge;
-import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.routing.vertextype.BikeRentalStationVertex;
-import org.opentripplanner.updater.GraphUpdaterManager;
-import org.opentripplanner.updater.GraphWriterRunnable;
-import org.opentripplanner.updater.JsonConfigurable;
-import org.opentripplanner.updater.PollingGraphUpdater;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import java.util.Map.Entry;
-import java.util.concurrent.ExecutionException;
+
+import static org.opentripplanner.graph_builder.linking.StreetSplitter.DESTRUCTIVE_SPLIT;
+import static org.opentripplanner.graph_builder.linking.StreetSplitter.NON_DESTRUCTIVE_SPLIT;
 
 /**
  * Dynamic bike-rental station updater which updates the Graph with bike rental stations from one BikeRentalDataSource.
@@ -37,11 +45,13 @@ public class BikeRentalUpdater extends PollingGraphUpdater {
 
     private static final String DEFAULT_NETWORK_LIST = "default";
 
-    Map<BikeRentalStation, BikeRentalStationVertex> verticesByStation = new HashMap<BikeRentalStation, BikeRentalStationVertex>();
+    Map<BikeRentalStation, BikeRentalStationVertex> verticesByStation = new HashMap<>();
 
     private BikeRentalDataSource source;
 
-    private SimpleStreetSplitter linker;
+    private Graph graph;
+
+    private StreetSplitter splitter;
 
     private BikeRentalStationService service;
 
@@ -118,8 +128,8 @@ public class BikeRentalUpdater extends PollingGraphUpdater {
 
     @Override
     public void setup(Graph graph) throws InterruptedException, ExecutionException {
-        // Creation of network linker library will not modify the graph
-        linker = new SimpleStreetSplitter(graph);
+        splitter = graph.streetIndex.getStreetSplitter();
+
         // Adding a bike rental station service needs a graph writer runnable
         service = graph.getService(BikeRentalStationService.class, true);
     }
@@ -155,6 +165,7 @@ public class BikeRentalUpdater extends PollingGraphUpdater {
             // Apply stations to graph
             Set<BikeRentalStation> stationSet = new HashSet<>();
             Set<String> defaultNetworks = new HashSet<>(Arrays.asList(network));
+            LOG.info("Updating {} rental bike stations.", stations.size());
             /* add any new stations and update bike counts for existing stations */
             for (BikeRentalStation station : stations) {
                 if (station.networks == null) {
@@ -166,7 +177,7 @@ public class BikeRentalUpdater extends PollingGraphUpdater {
                 BikeRentalStationVertex vertex = verticesByStation.get(station);
                 if (vertex == null) {
                     vertex = new BikeRentalStationVertex(graph, station);
-                    if (!linker.link(vertex)) {
+                    if (!splitter.linkToClosestWalkableEdge(vertex, NON_DESTRUCTIVE_SPLIT, true)) {
                         // the toString includes the text "Bike rental station"
                         LOG.warn("{} not near any streets; it will not be usable.", station);
                     }
@@ -185,13 +196,28 @@ public class BikeRentalUpdater extends PollingGraphUpdater {
                 BikeRentalStation station = entry.getKey();
                 if (stationSet.contains(station))
                     continue;
-                BikeRentalStationVertex vertex = entry.getValue();
-                if (graph.containsVertex(vertex)) {
-                    graph.removeVertexAndEdges(vertex);
+                BikeRentalStationVertex bikeRentalStationVertex = entry.getValue();
+
+                // before removing the bikeRentalStationVertex, first find and remove all associated
+                // SemiPermanentSplitterVertices
+                for (Edge edge : bikeRentalStationVertex.getOutgoing()) {
+                    if (edge instanceof StreetBikeRentalLink) {
+                        StreetBikeRentalLink toStreetLink = (StreetBikeRentalLink) edge;
+                        StreetVertex streetVertex = (StreetVertex) toStreetLink.getToVertex();
+                        if (streetVertex != null && streetVertex instanceof SemiPermanentSplitterVertex) {
+                            splitter.removeSemiPermanentVerticesAndEdges((SemiPermanentSplitterVertex) streetVertex);
+                        }
+                    }
                 }
+
+                // remove the bikeRentalStationVertex from the graph if it's in there (why wouldn't it be?)
+                if (graph.containsVertex(bikeRentalStationVertex)) {
+                    graph.removeVertexAndEdges(bikeRentalStationVertex);
+                }
+
+                // first get the outgoing
                 toRemove.add(station);
                 service.removeBikeRentalStation(station);
-                // TODO: need to unsplit any streets that were split
             }
             for (BikeRentalStation station : toRemove) {
                 // post-iteration removal to avoid concurrent modification