+++ src/test/java/org/opentripplanner/graph_builder/module/FakeGraph.java
@@ -1,11 +1,11 @@
 package org.opentripplanner.graph_builder.module;
 
-import org.opentripplanner.model.FeedScopedId;
-import org.opentripplanner.model.Stop;
-import org.opentripplanner.graph_builder.linking.SimpleStreetSplitter;
+import org.opentripplanner.graph_builder.linking.StreetSplitter;
 import org.opentripplanner.graph_builder.model.GtfsBundle;
 import org.opentripplanner.graph_builder.module.osm.DefaultWayPropertySetSource;
 import org.opentripplanner.graph_builder.module.osm.OpenStreetMapModule;
+import org.opentripplanner.model.FeedScopedId;
+import org.opentripplanner.model.Stop;
 import org.opentripplanner.openstreetmap.impl.AnyFileBasedOpenStreetMapProviderImpl;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.vertextype.TransitStop;
@@ -130,10 +130,15 @@ public class FakeGraph {
         }
     }
 
-    /** link the stops in the graph */
-    public static void link (Graph g) {
-        SimpleStreetSplitter linker = new SimpleStreetSplitter(g);
-        linker.link();
+    /**
+     * Index the graph and then link all stations in the graph.
+     */
+    public static void indexGraphAndLinkStations (Graph g) {
+        // recreate a new streetIndex even if one exists to ensure all indexes are built. It is assumed that all tests
+        // that use this method do not need any previously street indexes created during building the graph
+        g.index(true);
+        StreetSplitter linker = (StreetSplitter) g.streetIndex.getStreetSplitter();
+        linker.linkAllStationsToGraph();
     }
 
 }