+++ docs/Changelog.md
@@ -5,6 +5,8 @@
 - Add Way Property Set for the UK (#2818)
 - Fixes surefire test failure during build (#2816)
 - Improve documentation for `mode` routing parameter (#2809)
+- Refactor StreetSplitter (#2758)
+- Refactor bike rental edge creation (#2787)
 
 ## 1.4 (2019-07-30)